BPF optimizer: Coalesce jumps to identical return statements to minimize size.
This adds an optimization pass which looks for jumps that go to return
statements that return the same value, and looks for opportunities to
make them jump to as few of these return statements as possible. This makes
the other return statements safely removable, which in turn minimizes the
program and allows more unconditional => conditional jump translations.

This optimization is particularly useful in syscall filter programs, which
tend to have a structure like follows (either through linear or binary
search):

```
	.. if (foo) goto A else goto B
	A: return rejected
	B: if (bar) goto C else goto D
	C: return rejected
	D: if (baz) goto E else goto F
	E: return rejected
	F: return accepted
	...
	(Another set of rules in the program):
	.. if (foo2) goto G else goto H
	G: return accepted
	H: if (bar2) goto I else goto J
	I: return accepted
	J: return rejected
```

After this optimization pass, this can be reduced to:

```
	.. if (foo) goto J else goto B
	B: if (bar) goto J else goto D
	D: if (baz) goto J else goto I
	...
	.. if (foo2) goto I else goto H
	H: if (bar2) goto I else goto J
	I: return accepted
	J: return rejected
```

PiperOrigin-RevId: 568636445