Prefer routes with local addr on outgoing NIC only for local traffic
Currently gVisor's routing logic always prefers routes that use a local
address that is assigned to the outgoing interface. This preference is
applied even for forwarded traffic, where the local address is left
unspecified on route lookup because the source address of the packet
belongs to some other node as opposed to the stack itself. This means
that when forwarding incoming traffic, the netstack will prefer routes
that go through a NIC with a local address endpoint over those that do
not, even if a better route exists.

Change this logic such that the preference for routes with a local
address assigned to the outgoing interface only applies for locally-
generated traffic.

PiperOrigin-RevId: 571023083